requirements.txt
flask
openai
replicate
python-dotenv
google-cloud-storage
google-cloud-firestore
requests
gunicorn
werkzeug
